fixed many bugs in DQN that caused the AI to not learn at all.  major one was that it would populate memory with terminal states
added some code for simulations before realizing the AI is deterministic so simulating is kind of dumb... but can simulate against random?  just as a benchmark.  todo.

also started rewriting play.py so that it doesn't have the same code over and over.  it's kind of a mess now and might break, todo.

wrote a tictactoe dqn for testing.